feat(iam, app): request fee from user and pass to smart contract
